refactor(react): resolve circular deps
