+++ packages/react/src/context/Toast/context.tsx
@@ -10,7 +10,9 @@ import {
   type PropsWithChildren,
 } from 'react'
 import type { ToastSentiment } from '@pluralsight/headless-styles/types'
-import { Portal, Show, Toast, ToastHeading, ToastText } from '../../index.ts'
+import { Toast, ToastHeading, ToastText } from '../../components/toast.tsx'
+import { Portal } from '../../components/Portal.tsx'
+import { Show } from '../../components/Show.tsx'
 import { DISMISS, SHOW, defaultDuration, toastReducer } from './reducer.ts'
 import type { ToastContextProps, ToastProps } from './types.ts'
